+++ library/hls/src/main/java/com/google/android/exoplayer2/source/hls/HlsSampleStreamWrapper.java
@@ -524,6 +524,11 @@ import java.util.Arrays;
     return true;
   }
 
+  @Override
+  public void reevaluateBuffer(long positionUs) {
+    // Do nothing.
+  }
+
   // Loader.Callback implementation.
 
   @Override